Functions

Functions

Below you will find all methods exported by Helpers.jl.

Index

Helpers.absdiff

Helpers.comm

Helpers.compare

Helpers.compare_full

Helpers.docommute

Helpers.effreldiff

Helpers.h5delete

Helpers.h5dump

Helpers.h5repack

Helpers.loadrng

Helpers.meshgrid

Helpers.meshgrid

Helpers.meshgrid

Helpers.reldiff

Helpers.restorerng

Helpers.saverng

Helpers.setrng

Helpers.sparsejl2py

Helpers.sparsepy2jl

Helpers.sparsity

Helpers.swapcols!

Helpers.swaprows!

Docs

Helpers.absdiff — Method.

absdiff(A, B)

Difference of absolute values of A and B.

source

Helpers.comm — Method.

comm(A, B)

Commutator AB - BA.

source

Helpers.compare — Method.

compare(A, B)

Compares two matrices A and B, prints out the maximal absolute and relative differences and returns a boolean indicating wether isapprox(A,B).

source

Helpers.compare_full — Method.

compare_full(A, B)

Compares two matrices A and B, prints out all absolute and relative differences and returns a boolean indicating wether isapprox(A,B).

source

Helpers.docommute — Method.

docommute(A, B)

Checks if the matrices do (approx.) commute.

source

Helpers.effreldiff — Function.

effreldiff(A, B, threshold=1e-14)

Same as reldiff(A,B) but with all elements set to zero where corresponding element of absdiff(A,B) is smaller than threshold. This is useful in avoiding artificially large relative errors.

source

Helpers.h5delete — Method.

h5delete(filename, element)

Deletes a group or dataset from a HDF5 file. However, due to the HDF5 standard it might be that space is not freed.

source

Helpers.h5dump — Function.

h5dump(filename)

Dumps the group/data tree of a HDF5 file.

source

Helpers.h5repack — Method.

h5repack(src, trg)

Repacks a HDF5 file e.g. to free unused space. If src == trgWrapper to external h5repack application.

source

Helpers.loadrng — Method.

loadrng(filename [; group="GLOBAL_RNG"]) -> MersenneTwister loadrng(f::HDF5.HDF5File [; group="GLOBAL_RNG"]) -> MersenneTwister

Loads a random generator from HDF5.

source

Helpers.meshgrid — Method.

meshgrid(xvec) = meshgrid(xvec, xvec)

Produces a 2D meshgrid X,X by repeating xvec in y-dimension and xvec in x-dimension.

source

Helpers.meshgrid — Method.

meshgrid(xvec, yvec, yvec)

Produces a 3D meshgrid X,Y,Z by repeating the input vectors.

source

Helpers.meshgrid — Method.

meshgrid(xvec, yvec)

Produces a 2D meshgrid X,Y by repeating xvec in y-dimension and yvec in x-dimension.

source

Helpers.reldiff — Method.

reldiff(A, B)

Relative difference of absolute values of A and B defined as

$\operatorname{reldiff} = 2 \dfrac{\operatorname{abs}(A - B)}{\operatorname{abs}(A+B)}.$

source

Helpers.restorerng — Method.

restorerng(filename [; group="GLOBAL_RNG"]) -> Void restorerng(f::HDF5.HDF5File [; group="GLOBAL_RNG"]) -> Void

Restores a state of Julia's random generator (Base.Random.GLOBAL_RNG) from HDF5.

source

Helpers.saverng — Function.

saverng(filename [, rng::MersenneTwister; group="GLOBAL_RNG"]) saverng(HDF5.HDF5File [, rng::MersenneTwister; group="GLOBAL_RNG"])

Saves the current state of Julia's random generator (Base.Random.GLOBAL_RNG) to HDF5.

source

Helpers.setrng — Method.

setrng(rng)

Replaces current Base.Random.GLOBAL_RNG with rng.

source

Helpers.sparsejl2py — Method.

sparsejl2py(X)

Convert julia sparse matrix to python (scipy csc) sparse matrix.

source

Helpers.sparsepy2jl — Method.

sparsepy2jl(X)

Convert python sparse matrix to julia sparse matrix.

source

Helpers.sparsity — Method.

sparsity(A)

Calculates the sparsity of the given array. The sparsity is defined as number of zero-valued elements divided by total number of elements.

source

Helpers.swapcols! — Method.

swapcols!(X, i, j)

Swaps cols i and j of X.

source

Helpers.swaprows! — Method.

swaprows!(X, i, j)

Swaps rows i and j of X.

source